Implement url, summary and weight components [ci]
Fixes #1531
