+++ apps/sv/frontend/src/routes/createProposal.tsx
@@ -4,20 +4,7 @@
 import { useSearchParams } from 'react-router-dom';
 import { createProposalActions } from '../utils/governance';
 import { SelectAction } from '../components/forms/SelectAction';
-import {
-  Box,
-  FormControlLabel,
-  Paper,
-  Radio,
-  RadioGroup,
-  TextField,
-  Typography,
-} from '@mui/material';
-import { useForm } from '@tanstack/react-form';
-import { DesktopDateTimePicker } from '@mui/x-date-pickers';
-import dayjs from 'dayjs';
-import { dateTimeFormatISO } from '@lfdecentralizedtrust/splice-common-frontend-utils';
-import { useState } from 'react';
+import { CreateProposalForm } from '../components/forms/CreateProposalform';
 
 export const CreateProposal: React.FC = () => {
   const [searchParams, _] = useSearchParams();
@@ -28,174 +15,4 @@ export const CreateProposal: React.FC = () => {
   } else {
     return <SelectAction />;
   }
-};
-
-type CreateProposalAction = (typeof createProposalActions)[number];
-
-interface CreateProposalFormProps {
-  action: CreateProposalAction;
-}
-
-const CreateProposalForm: React.FC<CreateProposalFormProps> = ({ action }) => {
-  const [searchParams, _] = useSearchParams();
-  const actionFromParams = searchParams.get('action');
-  const [effectivityType, setEffectivityType] = useState('custom');
-
-  const form = useForm({
-    defaultValues: {
-      action: action.name,
-      expiryDate: dayjs(),
-      effectiveDate: dayjs().add(1, 'day'),
-    },
-  });
-
-  if (actionFromParams !== action.value) {
-    return <Typography variant="h3">Invalid action selected: {actionFromParams}</Typography>;
-  }
-
-  return (
-    <>
-      <Box sx={{ mt: 10 }}>
-        <Paper
-          sx={{
-            bgcolor: 'background.paper',
-            p: 4,
-            display: 'flex',
-            justifyContent: 'center',
-            alignItems: 'center',
-          }}
-        >
-          <Box sx={{ minWidth: '80%' }}>
-            <form
-              onSubmit={e => {
-                e.preventDefault();
-                e.stopPropagation();
-                form.handleSubmit();
-              }}
-            >
-              <Box sx={{ display: 'flex', flexDirection: 'column', gap: 8 }}>
-                <Typography variant="h3">[Work In Progress]</Typography>
-                <form.Field
-                  name="action"
-                  children={field => {
-                    return (
-                      <>
-                        <Box>
-                          <Typography variant="h5" gutterBottom>
-                            Action
-                          </Typography>
-                          <TextField
-                            fullWidth
-                            variant="outlined"
-                            name={field.name}
-                            value={field.state.value}
-                            onBlur={field.handleBlur}
-                            disabled
-                          />
-                        </Box>
-                      </>
-                    );
-                  }}
-                />
-
-                <form.Field
-                  name="expiryDate"
-                  children={field => {
-                    return (
-                      <>
-                        <Box>
-                          <Typography variant="h5" gutterBottom>
-                            Vote Proposal Expiration
-                          </Typography>
-                          <Typography variant="body2" color="text.secondary" gutterBottom>
-                            This is the last day voters can vote on this proposal
-                          </Typography>
-                          <DesktopDateTimePicker
-                            value={field.state.value}
-                            format={dateTimeFormatISO}
-                            onChange={newDate => field.handleChange(newDate!)}
-                            slotProps={{
-                              textField: {
-                                fullWidth: true,
-                                variant: 'outlined',
-                              },
-                            }}
-                          />
-                        </Box>
-                      </>
-                    );
-                  }}
-                />
-
-                <form.Field
-                  name="effectiveDate"
-                  children={field => {
-                    return (
-                      <>
-                        <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-                          <Typography variant="h5" gutterBottom>
-                            Vote Proposal Effectivity
-                          </Typography>
-
-                          <RadioGroup
-                            value={effectivityType}
-                            onChange={e => setEffectivityType(e.target.value)}
-                          >
-                            <FormControlLabel
-                              value="custom"
-                              control={<Radio />}
-                              label={
-                                <Box>
-                                  <Typography>Custom</Typography>
-                                  <Typography variant="body2" color="text.secondary">
-                                    Select the date and time the proposal will take effect
-                                  </Typography>
-                                </Box>
-                              }
-                            />
-
-                            {effectivityType === 'custom' && (
-                              <DesktopDateTimePicker
-                                value={field.state.value}
-                                onChange={newDate => field.handleChange(newDate!)}
-                                sx={{
-                                  width: '100%',
-                                  mt: 1,
-                                  mb: 2,
-                                }}
-                                slotProps={{
-                                  textField: {
-                                    fullWidth: false,
-                                    variant: 'outlined',
-                                  },
-                                }}
-                              />
-                            )}
-
-                            <FormControlLabel
-                              value="threshold"
-                              control={<Radio />}
-                              label={
-                                <Box>
-                                  <Typography>Make effective at threshold</Typography>
-                                  <Typography variant="body2" color="text.secondary">
-                                    This will allow the vote proposal to take effect immediately
-                                    when 2/3 vote in favor
-                                  </Typography>
-                                </Box>
-                              }
-                            />
-                          </RadioGroup>
-                        </Box>
-                      </>
-                    );
-                  }}
-                />
-              </Box>
-            </form>
-          </Box>
-        </Paper>
-      </Box>
-    </>
-  );
 };
